Fix `make lint` and `make format`, enforce `biome check` in Nix/CI
The `make` commands were broken because we didn't have `biome` listed as
a dependency, and we didn't have a `biome.json` file.

There's unfortunately still some inconsistency here in that `nix fmt`
and `nix flake check` are going to use whatever version of biome is
present in `nixpkgs`, whereas `make lint` is going to use the version in
our `package-lock.json`. We could make this consistent, but I'm
inclined to wait for it to be a problem.