Add a build at `./dist/lib/@cubing/icons/legacy-inline-woff/`.
For information about why, see: https://github.com/thewca/worldcubeassociation.org/pull/10526

Note that this doesn't add
`@cubing/icons/css/legacy-inline-woff` to the
`"exports"` field in `package.json`. Since we only
support this build for the WCA and we know that
the WCA's use case disregards `"export"` field
semantics, this allows us to make it harder to use by accident.