Use nix *just* for tooling
We drop all the hackery necessary to get Nix to understand how to set up
a `node_modules` directory. Instead we just let bun go off to the
internet to do bun stuff. This isn't a "pure" as the old way of doing
things, but it's less surprising to bun developers.

This PR does *not* affect CI (we still get bun from whatever docker
actions we're using), although I would like to discuss changing that in
the future.