Specify artifact dependencies to avoid running unnecessary build work.
This still keeps `build-web` and the linting targets as `.PHONY` commands for simplicity and safety.

The dependencies are very fragile, and require significant discipline to maintain. Just to give one example: `./dist/lib/@cubing/icons/js/index.d.ts` depends on `./src/js/index.ts` instead of `./src/js/*.ts` or `./src/js/**/*.ts` because the latter are not portable. This means that if a new source file is added under `./src/js` it must be added manually. But it would also be easy to miss adding this file, because some development flows would be unaffected.


| Command                               | Before              | After               |
| ------------------------------------- | ------------------- | ------------------- |
| `hyperfile "make build-lib-js"`       | 340.9 ms ±  45.3 ms | 19.9 ms ±   1.2 ms  |
| `hyperfile "make lint"`               | 2.414 s ±  0.064 s  | 1.804 s ±  0.081 s  |
| `hyperfile "make clean build-lib-js"` | 373.9 ms ±  69.6 ms | 324.1 ms ±  10.7 ms |
| `hyperfile "make clean lint"`         | 2.393 s ±  0.033 s  | 2.402 s ±  0.037 s  |